Rename and move Result (#7736)
