+++ pytorch_lightning/trainer/connectors/logger_connector/result.py
